+++ design_motor.m
@@ -14,5 +14,5 @@ function motor = design_motor(x)
     motor.Cost = temp(7); 
     motor.Diam = temp(8) / 1000;
     motor.planArea=(pi/4) * motor.Diam^2;
-    motor.Num=x(4);
+    motor.Num=double(x(4));
 end
